feat(test): fix e2e go test cluster resources
